+++ PSIntro.psm1
@@ -2,10 +2,10 @@
 # 17 June 2025 - need accommodate culture detection problems on non-Windows systems Issue #6
 $culture = $PSUICulture ? $PSUICulture : "en-US"
 $baseDir = Join-Path -Path $PSScriptRoot -ChildPath $culture
-Write-Host "Imported strings from $baseDir"
-write-host "Detected culture: $culture"
-Import-LocalizedData -BindingVariable strings -UICulture $culture
-# -BaseDirectory $baseDir #-FileName PSIntro.psd1
+#Write-Host "Imported strings from $baseDir"
+#write-host "Detected culture: $culture"
+#need to account for InvariantCulture on non-Windows systems
+Import-LocalizedData -BindingVariable strings -BaseDirectory $baseDir -UICulture (Get-Culture).Name -FileName PSIntro.psd1
 
 Get-ChildItem -Path $PSScriptRoot\functions\*.ps1 |
 ForEach-Object { . $_.FullName }
@@ -31,7 +31,6 @@ $table = $PSStyle.Formatting.TableHeader
 $prompt = ((&prompt | Out-String) -replace "`n|`r", '').Trim()
 
 #determine the tutorial path
-
 $tutorialPath = Join-Path -Path $PSScriptRoot -ChildPath (Join-Path -path $culture -ChildPath 'tutorials')
 if (-Not (Test-Path -Path $tutorialPath)) {
     #use the en-US tutorials if the localized path does not exist