+++ CHANGELOG.md
@@ -4,7 +4,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.9.0] - 2025-06-17
+
+### Changed
+
+- Corrected casing on en-US folder which was affecting localization on non-Windows systems.
+
 ### Fixed
+
 - Changed tutorial prompt to use Y as the continuation character for all cultures.
 - Fixed localization issues on non-Windows systems where no culture is detected. [[Issue #6](https://github.com/jdhitsolutions/PSIntro/issues/6)]
 
@@ -70,7 +78,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release to GitHub
 
-[Unreleased]: https://github.com/jdhitsolutions/PSIntro/compare/v0.8.0..HEAD
+[Unreleased]: https://github.com/jdhitsolutions/PSIntro/compare/v0.9.0..HEAD
+[0.9.0]: https://github.com/jdhitsolutions/PSIntro/compare/v0.8.0..v0.9.0
 [0.8.0]: https://github.com/jdhitsolutions/PSIntro/compare/v0.7.0..v0.8.0
 [0.7.0]: https://github.com/jdhitsolutions/PSIntro/compare/v0.6.0..v0.7.0
 [0.6.0]: https://github.com/jdhitsolutions/PSIntro/compare/v0.5.0..v0.6.0